refactor: AuthContext 위치 auth로 이동
