+++ src/hooks/useAuth.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import AuthContext from "@/contexts/AuthContext";
+import AuthContext from "@/features/auth/contexts/AuthContext";
 
 export default function useAuth() {
   const auth = useContext(AuthContext);
